requirements.txt
python-telegram-bot==20.3
   requests==2.31.0
   sqlite3==2.6.0